+++ CHANGES.rst
@@ -9,7 +9,7 @@ New features:
 
 * Change the way tagged fields are managed in the protocol definitions
   (pr #1162 by @vmaurin)
-
+* Don't send more than one JoinGroup request when more than one Assignors are specified (pr #XXX by @ilolis)
 
 0.14.0 (2026-04-29)
 ===================